refactor: folder structure for core and tests directory
