+++ website/tests/test_badge_views.py
