+++ website/tests/test_main.py
@@ -17,7 +17,7 @@ from selenium.webdriver.common.by import By
 from selenium.webdriver.support import expected_conditions as EC
 from selenium.webdriver.support.ui import WebDriverWait
 
-from .models import (
+from website.models import (
     Activity,
     ContentType,
     Domain,
