+++ website/views/project.py
@@ -36,9 +36,9 @@ from PIL import Image, ImageDraw, ImageFont
 from rest_framework.views import APIView
 
 from website.bitcoin_utils import create_bacon_token
-from website.filters import ProjectRepoFilter
+from website.core.filters import ProjectRepoFilter
 from website.models import IP, BaconToken, Contribution, Contributor, ContributorStats, Organization, Project, Repo
-from website.utils import admin_required
+from website.core.utils import admin_required
 
 # logging.getLogger("matplotlib").setLevel(logging.ERROR)
